Introduce StoreObject Property Wrapper (#399)
This pull request introduces a new property wrapper, `StoreObject`. This
property wrapper is designed to instantiate and manage a `Store` object
for the lifecycle of the view without making the view reactive to
`Store` updates. This behavior differs from that of `@StateObject` with
`ObservableObject`, which causes the entire view to refresh when the
`Store` updates.

In many cases, it is undesirable for the entire view to refresh whenever
the `Store` updates. `Store` updates can be efficiently handled through
`StoreReader`, and with `StoreObject`, we can now manage the `Store`
object in sync with the view lifecycle without triggering view-wide
refreshes.

This new property wrapper provides a more granular control over the
reactivity of views to `Store` changes and can improve the efficiency of
views that use `Store` objects.