Fix Binding (#531)
It should not capture value for the getter of Binding. as someone will
capture the binding instance and modify the value from the binding.
If the binding was made from struct, the binder will set struct modified
using value from getter which is captured and might be old.
that causes the problem of rolling back the state.